OCI_Functions/python/http_utils_with_auth/requirements.txt
fdk>=0.1.91
requests
oracle-ads